Fixed another human error...

+++ app/src/main/java/com/tellago/models/Goal.kt
@@ -62,6 +62,8 @@ data class Goal(
     }
 
     fun delete() {
-        collection.document(gid).delete()
+        if (gid != null) {
+            collection.document(gid).delete()
+        }
     }
 }
